+++ src/worker_manager.py
@@ -14,7 +14,7 @@ from sc2.units import Units
 # from src.Manager import Manager
 # from src.ProductionQueue import ProductionQueue
 # from src.QueueItem import QueueItem
-from Manager import Manager
+from manager import Manager
 
 
 #  TODO: Mine, worker micro and mules, Building creation, Base relocation, Repair ?
